requirements.txt
matplotlib==3.6.3
numpy==1.24.1
Pillow==9.4.0
